Allow configuring bucket for profile shipping
Grafana has another product in the observability space  called Phlare
which can periodically scrape pprof profiles from applications, which
could be incredibly useful for debugging issues with Cilium and other
components. Like the rest of the LGTM stack, it requires object storage,
so this commit adds the option to create another bucket for storing
profiles.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>