Addons improve logic
This commit introduces a mechanism to improve
the installation process of EKS addons.
The kube-proxy addon is installed regardless of
the health of nodes. The coredns addon is
installed only if there are healthy nodes. Thus
the Cilium module shouldn't depend on the EKS
module for the installation oterhwise it will
cause a circular dependency. All default addons
can be disabled by setting the variable
`cluster_addons` to {}. Specific addons can be
enabled by providing the to be enabled addons
in the `cluster_addons` variable. Enabling
specific addons will disable the default ones
meaning they must be explicitly defined.

Signed-off-by: darox <maderdario@gmail.com>